[HYPRCTL COMMAND] Add 'vdm' command to interact with the VDM plugin
This commit introduces the helpers and implementation for the 'vdm'
command in the Hyprland VDM plugin.
The current implementation includes only a simple vdlist subcommand that
will list (now is a placeholder) the virtual desktops managed by the plugin.

To register a command, the helper function `registerCommand` is used, which takes
the command name, description, and a function pointer to the command's implementation.
The command implementation function receives the command arguments and returns
a string result.

Signed-off-by: Davide Cardillo <cardillo.davide85@gmail.com>