[COMMANDS] Manage hyprctl commands unregistration
Managed the unregistration of hyprctl commands to ensure that
commands are properly removed when no longer needed. This prevents
potential problems related to dangling pointers or memory leaks when
the plugin is unloaded and loads again.

To do this, a new array `registeredCommands` has been introduced to keep track
of all commnads to register. During the registration process, each new
pointer to a commnad is stored in m_vRegisteredCommands, used later for
unregistration.

The `unregisterCommands` function is called into the PLUGIN_EXIT function.

Signed-off-by: Davide Cardillo <cardillo.davide85@gmail.com>