[WORKSPACE] Add workspace_manager class and integrate it into the plugin
This call is a more generic of what it is need in this project. This
becouse it comes from another similar project that I am working on, and
that could be useful in the future to extend the functionalities of this
plugin.

This class handle all the workspace management, like creating,
deleting, renaming, moving workspaces and so on. It also handle the
mapping between workspace names and workspace IDs.

It's implementation uses the singleton pattern to have a single
instance of the class that can be accessed from anywhere in the code.

Signed-off-by: Davide Cardillo <cardillo.davide85@gmail.com>